pyproject.toml
[project]
name = "mc-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "google-genai>=1.47.0",
    "ipydatagrid>=1.4.0",
    "ipykernel>=7.0.1",
    "ipywidgets>=8.1.7",
    "jupyter-bbox-widget>=0.6.0",
    "matplotlib>=3.10.7",
    "mcio-ctrl",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pycocotools>=2.0.11",
    "pydantic>=2.12.3",
    "requests>=2.28.1",
    "ruamel-yaml>=0.18.16",
    "supervision>=0.26.1",
    "torch>=2.9.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.24.0",
    "tqdm>=4.66.5",
    "transformers>=4.57.1",
    "ultralytics>=8.3.222",
]

[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.uv.sources]
torch = [{index = "pytorch-cu130", marker = "sys_platform == 'linux'"}]
torchvision = [{index = "pytorch-cu130", marker = "sys_platform == 'linux'"}]
mcio-ctrl = { path = "../mcio_ctrl", editable = true }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "mypy>=1.13.0",
    "ruff>=0.8.1",
    "black>=24.10.0",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "isort>=5.13.2",
    "pip>=25.3",
    "ipdb>=0.13.13",
]
